requirements.txt
fastapi
uvicorn
pandas
numpy
httpx==0.28.1
python-dotenv
yfinance==0.2.63